[misc] move Large FAT32 and ext formatting to their own source
* Also add randomization to test images
