+++ .vs/rufus.vcxproj
@@ -337,6 +337,8 @@
     <ClCompile Include="..\src\drive.c" />
     <ClCompile Include="..\src\format.c" />
     <ClCompile Include="..\src\dos.c" />
+    <ClCompile Include="..\src\format_ext.c" />
+    <ClCompile Include="..\src\format_fat32.c" />
     <ClCompile Include="..\src\icon.c" />
     <ClCompile Include="..\src\iso.c" />
     <ClCompile Include="..\src\localization.c" />
